Kitchen Sink.
Started out small, and wrap the router-view in a v-container.
moved the marvel attributionHTML text to the topbar.

added "load more" to the character search results and the series results.
Tweaked image sizes.
Got rid of the tooltip from the images so they don't get bigger anymore.

tweaked the column sizes so that on small screens it's col12, but on med and larger its col6.