disable history mode.
web browser doesn't like it if you manually change the URL, even if it is
to a subdomain.

So if we are at http://host
and we manually put in http://host/series/256
the page does a full reload.
maybe if the app were placed in a subdomain already?
testing on "pile.onrender.com"